Revert "Add TabCrashRecoveryExtension and a feature flag for tab crash recovery"
This reverts commit 3336681b6a0b9f5c24a1efd6482ac73cffe9e2cf.
