+++ macOS/DuckDuckGo/Tab/Model/Tab.swift
@@ -57,7 +57,6 @@ protocol NewWindowPolicyDecisionMaker {
         var tunnelController: NetworkProtectionIPCTunnelController?
         var maliciousSiteDetector: MaliciousSiteDetecting
         var faviconManagement: FaviconManagement?
-        var featureFlagger: FeatureFlagger
     }
 
     fileprivate weak var delegate: TabDelegate?
@@ -272,7 +271,6 @@ protocol NewWindowPolicyDecisionMaker {
                 tab.delegate?.closeTab(tab)
             },
                           titlePublisher: _title.projectedValue.eraseToAnyPublisher(),
-                          errorPublisher: _error.projectedValue.eraseToAnyPublisher(),
                           userScriptsPublisher: userScriptsPublisher,
                           inheritedAttribution: parentTab?.adClickAttribution?.currentAttributionState,
                           userContentControllerFuture: userContentControllerPromise.future,
@@ -288,8 +286,7 @@ protocol NewWindowPolicyDecisionMaker {
                                                        certificateTrustEvaluator: certificateTrustEvaluator,
                                                        tunnelController: tunnelController,
                                                        maliciousSiteDetector: maliciousSiteDetector,
-                                                       faviconManagement: faviconManagement,
-                                                       featureFlagger: featureFlagger))
+                                                       faviconManagement: faviconManagement))
 
         super.init()
         tabGetter = { [weak self] in self }
@@ -304,15 +301,6 @@ protocol NewWindowPolicyDecisionMaker {
         if favicon == nil {
             extensions.favicons?.handleFavicon(oldValue: nil, error: error)
         }
-
-        tabCrashRecoveryCancellable = extensions.tabCrashRecovery?.tabCrashErrorPublisher
-            .receive(on: DispatchQueue.main)
-            .sink { [weak self] error in
-                guard let self, let url = content.urlForWebView else {
-                    return
-                }
-                loadErrorHTML(error, header: UserText.webProcessCrashPageHeader, forUnreachableURL: url, alternate: true)
-            }
 
         emailDidSignOutCancellable = NotificationCenter.default.publisher(for: .emailDidSignOut)
             .receive(on: DispatchQueue.main)
@@ -1048,7 +1036,6 @@ protocol NewWindowPolicyDecisionMaker {
     private var webViewCancellables = Set<AnyCancellable>()
     private var emailDidSignOutCancellable: AnyCancellable?
     private var faviconCancellable: AnyCancellable?
-    private var tabCrashRecoveryCancellable: AnyCancellable?
 
     private func setupWebView(shouldLoadInBackground: Bool) {
         webView.navigationDelegate = navigationDelegate
@@ -1302,6 +1289,41 @@ extension Tab/*: NavigationResponder*/ { // to be moved to Tab+Navigation.swift
         // when already displaying the error page and reload navigation fails again: don‘t navigate, just update page HTML
         let shouldPerformAlternateNavigation = navigation.url != webView.url || navigation.navigationAction.targetFrame?.url != .error
         loadErrorHTML(error, header: UserText.errorPageHeader, forUnreachableURL: url, alternate: shouldPerformAlternateNavigation)
+    }
+
+    @MainActor
+    func webContentProcessDidTerminate(with reason: WKProcessTerminationReason?) {
+        guard (error?.code.rawValue ?? WKError.Code.unknown.rawValue) != WKError.Code.webContentProcessTerminated.rawValue else { return }
+
+        let terminationReason = reason?.rawValue ?? -1
+
+        let error = WKError(.webContentProcessTerminated, userInfo: [
+            WKProcessTerminationReason.userInfoKey: terminationReason,
+            NSLocalizedDescriptionKey: UserText.webProcessCrashPageMessage,
+            NSUnderlyingErrorKey: NSError(domain: WKErrorDomain, code: terminationReason)
+        ])
+
+        let isInternalUser = internalUserDecider?.isInternalUser == true
+
+        if isInternalUser {
+            self.webView.reload()
+        } else {
+            if case.url(let url, _, _) = content {
+                self.error = error
+
+                loadErrorHTML(error, header: UserText.webProcessCrashPageHeader, forUnreachableURL: url, alternate: true)
+            }
+        }
+
+        Task {
+#if APPSTORE
+            let additionalParameters = [String: String]()
+#else
+            let additionalParameters = await SystemInfo.pixelParameters()
+#endif
+
+            PixelKit.fire(DebugEvent(GeneralPixel.webKitDidTerminate, error: error), frequency: .dailyAndStandard, withAdditionalParameters: additionalParameters)
+        }
     }
 
     @MainActor
