+++ macOS/DuckDuckGo/Tab/TabExtensions/TabExtensions.swift
@@ -76,7 +76,6 @@ protocol TabExtensionDependencies {
     var tunnelController: NetworkProtectionIPCTunnelController? { get }
     var maliciousSiteDetector: MaliciousSiteDetecting { get }
     var faviconManagement: FaviconManagement? { get }
-    var featureFlagger: FeatureFlagger { get }
 }
 
 // swiftlint:disable:next large_tuple
@@ -88,7 +87,6 @@ typealias TabExtensionsBuilderArguments = (
     setContent: (Tab.TabContent) -> Void,
     closeTab: () -> Void,
     titlePublisher: AnyPublisher<String?, Never>,
-    errorPublisher: AnyPublisher<WKError?, Never>,
     userScriptsPublisher: AnyPublisher<UserScripts?, Never>,
     inheritedAttribution: AdClickAttributionLogic.State?,
     userContentControllerFuture: Future<UserContentController, Never>,
@@ -226,15 +224,6 @@ extension TabExtensionsBuilder {
             FaviconsTabExtension(scriptsPublisher: userScripts.compactMap { $0 },
                                  contentPublisher: args.contentPublisher,
                                  faviconManagement: dependencies.faviconManagement)
-        }
-
-        add {
-            TabCrashRecoveryExtension(
-                featureFlagger: dependencies.featureFlagger,
-                contentPublisher: args.contentPublisher,
-                webViewPublisher: args.webViewFuture,
-                webViewErrorPublisher: args.errorPublisher
-            )
         }
 
 #if SPARKLE
